services/hair-tryOn-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2